RE-279 Build/save image in public cloud
In order to allow the capability to save an image
from a prepped instance in public cloud, a job
added which does so.

A custom python script is used due to the fact
that pyrax appears to be the only reliable way
to do the action in RAX PubCloud. The nova CLI
does not seem to work and we need to use pyrax
for the rax Ansible modules. This can be revisited
when we switch to using the shade-based modules
instead.