RE-197 dependency_update hook
This commit adds a standard job that can be used by all repos/projects
that want to keep dependencies up to date.

The interface is simple, the repo should add a script at
gating/dependency_update/run. If that script dirties the repo
when run, then a PR will be created with the changes.