RLM-204 Fix build failure issue creation
Currently the mechanism that creates a jira issue when a periodic
build fails is not working. The problem is a call:
  common.create_jira_issue(project="RO")

This looks like a calll that passes a keyword argument, its not. Its
an assignment with a function call. In CPS groovy assignments return
null, so this is effectively common.create_jira_issue(null).

The fix is to remove project=.