RE-451 Fix bug getting OSA SHA on multi-node
This changes updates how the OSA SHA is identified, for use by the
openstack-ansible-ops multi-nodes scripts, so that it accounts for use
of a submodule or being specified in a file. This change is required due
to the RPCO commit 4d12456c1ab1111823e7b0cb8f01d37a4b2a7f4f.