Prevent Unnecessary JJB Runs
Currently JJB runs whenever a commit is pushed to any branch in rcbops/rpc-gating. Everytime it runs it deploys master, so it should only run when master changes. 

This is especially annoying if you are testing a job from a different branch, as when you push to your branch the JJB job will be triggered and revert the job to its master state.