RLM-214 Leapfrog Gate Job Modifications
* Removes all kilo point release branch jobs from
AIO/Multi as we are not going to test test individual
point release of Kilo.
* Sets up K->N, K->N-14.2.0, L->N, L-> N-14.2.0 Periodics
as these will be our primary gate jobs to know if things broke.
Since we're getting ready for Liberty, adding these now as well.
* Moves leapfrogupgrade items to the _leapfrog config files, so
we can avoid issues with the existing jobs and own changes in
our config