RE-577 Improve respawn behaviour of webhook tunnel
CIT Jenkins is hosted internally with no forwarding for receiving
webhooks from github. To work around this we host a webhook proxy
in public cloud, this runs nginx, receives webhooks from github and
forwards them via an ssh tunnel to CIT Jenkins.

The ssh tunnel must be started on the Jenkins master and connect
out to the webhook proxy node as connections in the other direction
are not possible. The webhook setup job creates a script on the
Jenkins master that in turn creates an upstart job to manage the tunnel.

This commit improves the upstart job by removing the respawn limit
and adding a delay between respawn attempts.