RE-506 Implement pre_merge_test standard job
As documented in [1] this patch implements a standard
job for the pre_merge_test hook. It also implements
a job for the rpc-gating repository to execute it
which can replace the current Single-Use-Slave-Example
job.

The artifact publishing across all jobs now also uses
a common function - this removes the current duplication
and ensures consistency in the application. In order to
achieve this the collection of job artifacts for the
RPC-O jobs has been ranamed to be RPC-O spcific.

[1] https://rpc-openstack.atlassian.net/wiki/spaces/RE/pages/19005457/RE+for+Projects