RE-613 Skip artifact upload if dir not found
Currently IRR jobs are failing as they write logs to WORKSPACE/logs,
but the artifact upload playbook is expecting them in
WORKSPACE/artifacts. This commit implements the first part of the
solution which is to skip upload if WORKSPACE/artifacts doesn't
exist.

The second half is to modify IRR jobs to write artifacts into
WORKSPACE/artifacts so they are archive correctly.

This commit also removes artifact compression from the IRR job
as that is now handled in pubcloud.uploadToCloudFiles.