Update Ansible and adjust MNAIO setup
Ansible 2.2.2.x seems to have issues with ssh disconnects as noted here:
[Related-Issue]. Reports indicate going to 2.3.x resolves some of the issues.

We don't need to run the "openstack-service-setup.yml" playbook. We deploy
RPC-O in our own way, not following the method in the MNAIO, so this playbook
is not used by us.

SSH params are now being passed into the jenkins slave setup playbook. This is
done to ensure we're not causing ourselves issues with SSH configs as we setup
the Jenkins slave on the infra1 VM instead of the OnMetal Host; failures with
SSH disconnects and host key verifications have been repeatedly observed
in the gate.

To resolve an issue with the RPC-O extra inventory the `dynamic-inventory.py`
script has been converted to write a YAML based inventory instead of using
the JSON type which does not directly support the standard Ansible inventory
extensions.

Related-Issue: https://github.com/ansible/ansible/issues/25941
Issue: https://rpc-openstack.atlassian.net/browse/RE-632
Co-Authored-By: Matt Thompson <matt.thompson@rackspace.com>
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>