RE-674 Adjust apt sources if necessary
When setting up a single-use slave, we need to install
packages. Some of those packages may also be installed
later after apt sources are reconfigured (as is the case
with RPC-O), but the packages in those sources are older
and the package database therefore errors.

In this patch we try to detect whether we need to use
the apt artifacts as a source and configure them before
installing any packages. This hopes to prevent jobs
failures due to the packages we install for Jenkins.