RE-507 post merge job
This commit implements the standard job for post merge tests.
It will also be the basis for merge triggered jobs, though that will
likely require a separate free-style job that triggers this one.

This commit also includes an implementation of the post_merge_test hook
for rpc-gating. That will create a periodic job that runs the lint
tests. That job can be used as an example for other projects.

Also:
  * Increased delay before attempting to download the gating venv
    after starting a job to build one. It was 10s, that was
    insufficient, it is now 60s.
  * Add test result gathering to common.archive_artifacts.
  * Fix lint job to not rely on shell glob ordering.