RLM-219 Remove ClientAliveInterval from sshd role
Was having issues with some jobs timing out
and losing their connection to the slave.
Removing the ssh role seemed to fix the problem
so I tried the ClientAliveInterval and it seemed
to help correct the issue.

15 sec is probably too low for the server to hear
back from the client, so removing this to see if
it helps get rid of the
"hudson.remoting.RequestAbortedException:
java.io.IOException: Unexpected termination of the
channel" errors