RE-634 Change Merge-Trigger-JJB to freestyle job
Currently, Merge-Trigger-JJB fires irrespective of the branch on
rcbops/rpc-gating being pushed to.  We tried to circumvent this by
using GIT_BRANCH inside the pipeline script, however that environment
variable is not being passed in and therefore this job will never build
the Jenkins-Job-Builder job.

Since there is no clear way to get a pipeline script to only trigger
when a specific branch is being pushed to, we need to convert this job
to a freestyle job where we can specify the scm and specific branch(es)
we wish to build against (master, in this case).

NOTE: It's likely we'll need to delete the existing Merge-Trigger-JJB
      job when this merges, as I don't believe JJB will change project
      types on an existing job.