RE-731 Fix venv build file changed while reading
The venv build job has been failing due to a file changing while
the venv tar is being created. I think this is due to the job
using a dirty workspace. It shouldn't be due to parallelism as the
job doesn't allow parallelism. However to make that explicit I've added
a serialisation lock around the job.