RE-643 Phobos <--> Webhook Proxy Tunnel
As internal routing between CIT jenkins and the Phobos environment
is not possible in the short term, a tunnel is added.

The deploy-phobos container in the phobos environment will connect
to the webhook proxy node via ssh. The ssh connection will carry
a reverse port forward such that any connections to port 2223 on
the webhook proxy  node will be forwarded to port 22 on deploy-phobos.

Jenkins will ssh to webhookproxy:2223 when it requires a connection
to deploy-phobos.