RE-722 Use pre_merge_test hook for RPC-O mitaka
This implements the standard pre_merge_hook for the
RPC-O mitaka branch and excludes the corresponding
rpc_aio PR triggered jobs. The project name uses the
OpenStack series as there is no way to filter the
'trusty' image from the master branch using the
standard job at this time, nor is there a way to
filter the 'xenial' image from the newton project
if that were to be used.