RE-723 Use pre_merge_test hook for RPC-O liberty
This implements the standard pre_merge_hook for the
RPC-O mitaka branch and excludes the corresponding
rpc_aio PR triggered jobs. While the job could make
use of the 'mitaka' project as the settings are all
the same, its own project name was used to keep to
the pattern (it keeps the job names intuitive) and
to make it simple to remove at a later date.