RE-724 Use pre_merge_test hook for RPC-O kilo
This implements the standard pre_merge_hook for the
RPC-O mitaka branch and excludes the corresponding
rpc_aio PR triggered jobs. The job uses its own
project name was used to keep to the pattern (it
keeps the job names intuitive) and to make it simple
to remove at a later date.