Add selective PR trigger to pre_merge_test hook
The ironic scenario is still currently being developed
and is not yet ready for triggering on every PR. As
such there is the capability in the current job
definition to disable the test until specifically
triggered.

This patch implements the same functionality into the
pre_merge_test hook so that all jobs can make use of
it. It also adds the use of this to the new ironic
jobs.