RO-2796 Use workspace for apt artifact checkout
As the apt artifacts are built using a long-lived
node, the jenkins workspace should be used for the
git checkout the job does to prevent the constant
back-and-forth switching between branches from
interfering with each other.

The script being executed from rpc-openstack does
a symlink from the current directory to the expected
path of /opt/rpc-openstack so everything else should
continue to work as expected.