RE-759 Fix notification e-mail message
Issues:
The e-mail message body is set to the file path where the release notes
are saved. The command-line option `--body` is being incorrectly passed
the path instead of the actual text to use in the message.

In general, the sub-command `mailgun` should be run after
`generate_release_notes` and so the message body should not need to be
supplied however the click context object is not updated to include the
release notes and so the incorrectly supplied path is used instead.

Release notes should be formatted in GitHub Markdown, e-mail clients are
not guaranteed to present this in a reader-friendly format.

Change:
This change removes from `ghutils.create_release` the responsibility of
defining the message subject and consolidates all control of the e-mail
to the notifications file. To help to keep the separation of concerns,
`ghutils.create_release` now updates the context object to include the
link to the new release so it can be used in the message instead of
sending the release notes text.

`notifications.generate_message_data` is used to unify all the common
code, in `notifications.mail` and `notifications.mailgun`, that is being
used to inspect the context object and determine the sender, subject,
and body of the message.