RE-633 Update mailgun creds
I missed the fact that we already had a mailgun account setup, and had
DNS entries created for a domain created in a new mailgun account.
Rather than move the domain to the old mailgun account (which would
require updating the previously created DNS records), I've just added
new credentials in Jenkins for the new account.  We can remove the old
credentials or just leave them around in the event that we need to flip
over to another Mailgun account.