Implement optional job_name_prefix
It's common in complex job definitions to need to do
multiple projects for a single repo, but in doing so
the job names get very long because JJB projects have
to be unique.

In this patch we implement the optional 'job_name_prefix'
variable which can be set in order to allow the job name
to use a shorter prefix than the full project name.