RE-508 Remove pyrax
In preparation for using Phobos, the use of pyrax and rax Ansible
modules needs to be replaced with the generic OpenStack versions. The
Python modules have conflicting dependencies and therefore cannot both
be installed and working in the gating venv. This change removes pyrax
from the list of requirements and adds shade to enable use of the os_*
Ansible modules. rackspaceauth is required to use the os_* modules with
Rackspace public cloud.

These changes do not, by themselves, allow for Phobos to be used by
rpc-gating. Use of Phobos will need to be added separately once the
nature of its use is decided.

common.groovy

`writeCloudsCfg` is added to create a configuration file compatible with
the new os_* Ansible modules, it is used instead of `writePyraxCfg` to
make the necessary credentials available. `writeRaxmonCfg` is for use
by the raxmon CLI tool.

allocate_pubcloud.yml

This change updates allocation_pubcloud.yml to use the community
OpenStack Ansible modules instead of Rackspace-specific ones.

The switch from `rax` to `os_server` is mostly a case of modifying
parameter names. The exceptions to that are:
- `os_server` requires the section of clouds.yaml to be specified with
  the parameter `cloud`
- `os_server` does not support building multiple instances. Instead of
  attempting to retain this functionality in the playbook, it has been
  removed give that `count` is always set to 1.

cleanup_playbook.yml

Use of `rax` is replaced with `os_server`. Given that creating multiple
instances is no longer supported by allocate_pubcloud.yml this playbook
has been simplified to remove a single instance.

create_cloud_image.py

The script rax_create_image.py has been renamed to create_cloud_image.py
given that in is no longer tied to Rackspace public cloud. It has been
mostly rewritten due to the differences between the pyrax and openstack
APIs however, the suggestion in the comment at the top of the file to
switch to the Ansible modules has not been followed because it would
change how the image is updated. Changes beyond the switch away from
pyrax are being left to future refactoring to simplify this patch.

To use the openstacksdk configuration file, the script required a couple
of changes.  The file location is now specified using the environment
variable `OS_CLIENT_CONFIG_FILE` and the cloud to select from the file
is set using the option `--cloudname`.

upload_to_swift.yml

The Ansible module os_object does not support setting the object to be
automatically deleted by Swift after a defined period of time. The
`swift` CLI tool is used to upload the archive, to enable this to be
configured.

The public object URL for CloudFiles is determined via the CDN API, this
is not something supported by the OpenStack Swift tools, and so there is
now additional code to get the correct URL based on whether or not
CloudFiles is being used.

maasutils.py

pyrax is replaced here with rackspace_monitoring, this is a relatively
trivial change given that pyrax itself uses rackspace_monitoring to
support the monitoring API.

multi_node_aio_maas_entities.yml and multi_node_aio_maas_cleanup.yml

To remove pyrax, the CLI tool raxmon is called to create and delete the
entities. Given there is no longer a single entry point to these two
actions the role maas_entities is removed and the required tasks are
added to the playbooks.

periodic_cleanup.py

This change replaces the usage of pyrax with openstacksdk.  At present
the script does not handle command-line arguments, so to keep changes at
a minimum we hard-code the clouds.yaml cloud to be 'public_cloud'.

Co-Authored-By: Matt Thompson <mattt@defunct.ca>