RE-617 Update standard_job_premerge_release.yml
This job template seems problematic in that jobs are currently failing
with:

```
+ python rpc-gating/scripts/release.py
Usage: release.py [OPTIONS] COMMAND1 [ARGS]... [COMMAND2 [ARGS]...]...

Options:
  --org TEXT            Github Organisation that owns the target repo
                        [required]
  --repo TEXT           Name of target repo  [required]
  --pat TEXT            Github Personal Access Token  [required]
  --debug / --no-debug
  --help                Show this message and exit.

Commands:
  add_issue_url_to_pr
  clone
  create_issue
  create_pr
  create_release
  generate_release_notes
  mail                    Send mail via local MTA
  mailgun                 Send mail via mailgun api.
  publish_tag
  update_rc_branch        Update rc branch.
  usage
+ --debug
```

While the `COMMAND` parameter YAML looks correct (>- to unfold the
lines, removing newline at the end), viewing the long `COMMAND` string
in the Jenkins UI is difficult when the field is a `string`.

This commit simply updates `COMMAND` to be a text, changes the >- to
|, and appends each line with a slash.  This is the format currently
used in rpc_jobs/release.yml.