RE-1019 Fix fallback region split
Commas in the fallback region string are replaced with spaces, but then
the string is split on commas to form a list. Consequently the string
is not split and a single item list is produced. This list is not a
valid region, so attempts to use the fallback region fail.

This commit changes the allocate_pubcloud playbook to split the
fallback region list on spaces rather than commas to avoid the above
problem.