RE-1020 Introduce protected_prefix
Previously instance_prefix was used to match instances that should
be cleaned up. The idea was that all jobs used a common prefix
(like rpc-aio) and we could use that prefix to identify temporary
nodes that had leaked.

However there are now too many jobs to accurately maintain the
prefix regex. It also failed to account for test jobs that could
create instances that didn't match the prefix.

This commit inverts the logic, removes instance_prefix and replaces
it with protected_prefix. All instances that don't match the
protected_prefix regex, will be subject to cleanup.