RE-1046 Bump openstack-ansible-ops
This is necessary to work around failures we're hitting as a result of
ansible 2.4.1.0 being installed.

Note that in addition to bumping the SHA, we also need to introduce a
new host var, `server_domain_name`.

Lastly, because openstack-ansible-ops now reverts to using ansible
2.3.2.0, we have to revert some of the python ansible code in
pipeline_steps/maas.groovy.  However, instead of reverting wholesale we
update it to work with both ansible < 2.4 and >= 2.4.