RE-1073 Check for inventory file before clean up
The clean up process requires an inventory file, in order
to determine the ID of the instance to remove and the region that
contains it.

If the inventory file doesn't exist, there is no point attempting
cleanup. If the job failed or was aborted before instance allocation
then there is nothing to cleanup. If an instance was allocated then
it will be cleaned up by the periodic cleanup job.