RE-809 Remove reference to wherenow.org
RE now has access to the rpc.rackspace.com DNS domain, we therefore
do not need to use personal domains such as wherenow.org for anything.
This commit removes a reference to wherenow.org. The new domain
points to the same IP, so re-deploys will work as the do currently.