RE-652 Correct stashing
Currently, we're using `include` when we stash instead of `includes`,
and this is including far too many files.

This commit simply corrects each stash to refer to `includes` instead.
Note that in multi_node_aio_prepare.groovy, we also update the stash to
point to the specific inventory file relative to the working directory,
rather than referring to `inventory_path`, which didn't seem to work.