RE-1082 Use alt rpc-maas dir for phobos verify
Currently phobos verification uses /opt/rpc-maas however that is
modified by the rpc-openstack deployment tools, so a different version
may be used before and after deploy. It also conflicts with the
rpc maas branch parameter, and potentially makes debugging hard
as the same repo may be at 3 different versions during a job.

This commit introduces a separate directory for the version of
rpc-maas used by phobos verify. This also requires a commit in
rpc-eng-ops to read the playbook from the new location.