RE-1098 Dont create issues for timeout exceptions
Ideally timeouts would be within the node block so that clenaup
cannot fail due to timeout. However that is more complicated
when the node block is reused in different situations (eg
runonpubcloud).
As we are planning to move to node pool, its not worth investing
the time to refactor that, so drop timeout exceptions instead
of creating useless jira issues.

As this requires referring to an exception in pubcloud.groovy, that
exception has to be available during lint, which it isn't as lint
is not run within Jenkins. To get around this, I implemented a dummy
exception class as was done previously with NonCPS. However I realised
that all the scripts in pipeline_steps are loaded in jobs and didn't
want the potential for a conflict with the real class, so moved the
lint dummy classes into a different folder.