RE-678 Add merge trigger job for post merge job
Currently the post merge standard job is triggered by cron/timer only.
This commit adds a freestyle job which is triggered by pushes to github
which in turn triggers the post merge standard job.

The reason for this is that github push trigger is not reliable with
pipeline jobs (unless using orgfolder/multibranch pipeline which we
aren't).

Each project that wants to use a push trigger, will create a JJB
project to instantiate the job for their repo. Re-For-Projects
will be updated with instructions.