RE-1114 Remove generator usage from cleanup script
Generators can only be iterated once, and in one case, we
were doing two membership tests on one generator, causing incorrect
results for the second test.

This commit removes all generator usage, even when it wouldn't
cause a problem, as the data set is small enough that we don't need
the optimisation. This may also prevent bugs being introduced in future.