RE-35 Adjust save_pubcloud for fallback regions
9ade8a7e95648a2c3e664a0fe5db41d4de5e9bb1 Introduced fallback regions,
so if an instance fails to build it can be built in an alternate
region. This means the region that will be used is not known
before the job starts. The region is written to an inventory
file when the instance is created.

In that commit save_pubcloud was modified to use rax_region from
the inventory file, but the hosts pattern was not modified to use
the node that the variables are set for.

This commit adjusts the hosts pattern and delegate statements to match.