RE-35 Execute thaw hook when creating instances
When an instance is created, check for the presence of /gating/thaw
and execute it if present. This early hook allows for instance
reconfiguration before the Jenkins slave is connected.

If there is a requirement to make changes earlier in the boot process,
the file system can be modified before the snapshot is created.
The best hook for that is gating/post_merge_test/post.