RE-53 Add lint check to validate job name length
We've run into situations where long job names cause python scripts in
virtualenvs to fail.  This is due to a kernel limitation that imposes
the max number of characters that can exist in an interpreter line in a
script.

This commit adds a new check that scans all job name lengths and then
errors if any exceeds 77 chars.