RE-1126 Update when in server provisioning task
The when clause in the 'Provision a cloud instance' task was not updated
in the migration from rax module to os_server, and this is resulting in
an instance being created in all regions, rather than just the first.

This commit updates the when clause to work with os_server.

Co-Authored-By: git-harry <git-harry@live.co.uk>