RE-1139 Fix retry instance creation
The task "Provision a cloud instance" fails without trying the next
region under certain types of build failure, for example if the
requested image does not exist in the current region.

This change updates the when clause so that the next region will be
tried if no 'server' key is present in the result object.