RE-774 Add withRequestedCredentials
This allows jobs to request credentials while only knowing the
credential ID. This is an improvement over the usual mechanism
as that requires the user to specify id, type and the names
of environment variables to populate.

This change allows credentials to be specified via JJB. That isn't
practical using withCredentials as that requires a list of dictionaries
which can't be encoded into a string variable in a user friendly way.