RE-1136 Override REGIONS in rpc_upgardes.yml
Currently we're inherting from defaults.yml, which only has a single
region in REGIONs.  This means builds will fail when that region is
full (which it is now).

Note that the onmetal flavour we're using is only available in IAD,
so we override REGIONS w/ just that region for mnaio builds.