RE-1129 Implement nodepool setup job
Currently most RPC tests use single-use slaves which are requested,
built, used and deleted within the test time of every job. This adds
a minimum of 10 minutes of test time to every job and often results
in failures. This mechanism also adds a bunch of management code and
administration which we can displace quite easily by using nodepool.

Nodepool is what OpenStack uses to prepare images and cloud instances
to test on. It handles the preparation and teardown so that the tests
only have to request and release the nodes.

This patch adds a job to do the build of nodepool hosts to manage the
system. The intent for this job is to continuously deliver the software
and configuration to the instances running as nodepool servers.

For now this only handles the initial setup. Follow on work will add
the jenkins enablement, monitoring and the image preparation.

Included in this patch are also the following changes:

* A new groovy function is added to use the rpc-gating venv to
  execute ansible-galaxy to download the contents of a standard
  ansible-role-requirements.yml file.

* The clouds.yaml configuration implemented is changed to include
  options to allow the file to be used along with the Ansible
  OpenStack dynamic inventory. The regions are also specified
  in clouds.yaml as that is required when using the dynamic
  inventory.

* The drop_ssh_auth_keys playbook is adjusted to allow it to target
  a given set of hosts. The default remains 'job_nodes' to ensure
  that it does not break anything else using it. The concatenation
  for the file lookup is adjusted to ensure that it does a string
  concatenation - this is just good practise.

* The openstack dynamic inventory script for Ansible 2.3.2.0-1
  is vendored into the git tree as it is the most reliable way
  of accessing it to use it. As the script is a contrib script,
  it is unfortunately not included in the pip package for ansible.

* python-novaclient is added to requirements.txt in order to
  support the os_server_actions ansible module. The entry for
  python-swiftclient is moved so that the file's entries are
  sorted alphabetically.

* The slave_open_ports.yml playbook has been consolidated into
  the slave_security.yml playbook as that was the only place it
  was called from and there is no need for it to be a seperate
  playbook.

* The slave_security playbook is adjusted to allow it to target
  a given set of hosts. The default remains 'job_nodes' to ensure
  that it does not break anything else using it. It now also
  implements sshd configuration to improve the security posture
  for any instances it executes against.

* The jenkins slave setup is adjusted to allow sshd to listen on
  all interfaces (including ipv6). This is done so that it does
  not clash with the slave_security settings. Connectivity via
  ipv6 is ansible's default, so it's necessary for that purpose.
  In subsequent work we can cut out the duplicate execution of
  the sshd configuration role, but it is out of scope for this
  work.