RE-787 Add PR build skipping for matching files
This patch introduces the ability in a standard job to define a skip
regex. If all the paths of all the files modified by the pull request
match the regex, the build is ended and marked a success. This allows
builds to be skipped when they aren't impacted by the pull request
changes.

The regex support is provided by Python's re module with the verbose
flag enabled [1], this allow for more readable multi-line regex to be
written without the whitespace impacted the expression. When defining
`skip_pattern` in YAML use `>-` so that the string does not include new
lines.

This change also adds a regex for rpc-openstack pull requests to fix the
functionality lost when those jobs were moved from using rpc_aio.yml to
standard_job_premerge.yml.

This facility can be applied to the project, or to an individual axis.
The skip regex is also entirely optional, by default no build is
skipped.

`common.is_doc_update_pr`, the original implementation of this
functionality has been left untouched. It can be removed once all
old-style jobs using it have been removed.

[1] https://docs.python.org/2/library/re.html#re.VERBOSE

Co-Authored-By: git-harry <git-harry@live.co.uk>