RE-35 Add trigger detection
This commit adds setTriggerVars to common. This function
detects the cause of the current build and sets RE_JOB_TRIGGER
and RE_JOB_TRIGGER_DETAIL appropriately. Note that not all
triggers have additional detail for the second var, in that
case it will be "No detail available".

The detected triggers are:
* UPSTREAM: another job
* USER: manually initiated
* TIMER: cron / periodic
* PUSH: A commit was pushed to a repo
* PULL: Pull Request
* OTHER: Unknown trigger

RE_JOB_TRIGGER is a previously published env var as part of the
standard jobs interface. There is only one usage, which is in the
rpc-octavia project, which will be adjusted to match this change.