RE-35 Use Standard 15 for RPC post merge jobs
The post merge jobs will be used for saving snapshots. However if a
flavour with ephemeral storage is used, the snapshot will only
contain the base OS, not the containers.

To prevent this, we use a flavour with a large system disk and no
ephemeral storage so that the whole deployment is stored in the
snapshot (7 / Standard 15)

This commit also adds post merge trigger jobs for RPCO, as they
didn't exist previously.