RE-1151 Add playbook / jobs to deploy repo servers
This commit leverages the work done in [1] to continuously deploy
repo servers for artifact storage.

In order to do this, setup_openstack_instances.yml had to be
modified to provision block storage, which is required for storage of
artifacts.  This primarily involves:

- adjusting the initial unlock to only happen when the instance is
  marked for deletion (state=absent), or when there are volumes defined
  in instance_list
- adding the necessary tasks to deploy and attach the specified volumes
- adding an inventory refresh to provide an updated inventory for when
  we need to query inventory to determine if an instance needs to be
  locked
- finally, we check if the instance is in the inventory to determine if
  it needs to be locked

Note that marking an instance as state=absent does not result in the
instance's volumes being torn down.  This is something we can focus on
in a later iteration if we find we need to automatically tear down
volumes.

TODO:
  - lock down any necessary ports w/ ufw
  - look at combining setup_openstack_instances.yml,
    drop_ssh_auth_keys.yml, and slave_security.yml into the same
    venvPlaybook execution in both rpc_jobs/setup_nodepool.yml and
    rpc_jobs/repo_server.yml (this will avoid an extra inventory
    lookup since the inventory is refreshed at the end of
    playbooks/setup_openstack_instances.yml)

[1] https://github.com/rcbops/rpc-gating/pull/542/