Handle teardown of volumes on deleted instances
This commit updates setup_openstack_instances.yml to handle the
teardown of volumes if they're specified as `state: absent` and are
associated with an instance that also has `state: absent`.

Currently we do not handle the teardown/deletion of volumes associated
with a running instance as we'd need to do some poking around inside
the instance to verify if the volumes are in a safe state to be
removed.  If this is an absolute requirement we can factor this in at a
later date.