RE-1203 Fix use of skip_pattern in job definitions
This change moves `skip_pattern` to being a parameter. This means that
it is no longer defined in the pipeline script generated from the `dsl`
section of the pre-merge job template. Making this change means that the
string definition in the project no longer needs to escape special
characters.

Making `skip_pattern` a parameter fixes the rpc-openstack pull request
build failures caused by the use of unescaped `"\"` and `"$"` in the
definition of `skip_pattern` for its projects.

The regex is now supplied to the search function using an environment
variable instead of interpolating the string into the script. This
prevents any issues with interpolating a bad string.

The appearance of the regex in the console log is improved by moving it
to a new line.

Switching to use a parameter means that a trailing new line is no longer
a concern when defining `skip_pattern`. While leaving them would cause
no harm, this change removes the block chomping indicator, `-`, to
prevent any confusion from unnecessary stripping.

The block scalar used in the project definitions is switched to `|` to
improve how the pattern is displayed in the log.