RE-1200 override MaaS Verify retries
At the end of the RE-1200 deploy, a large spike in rabbitmq messages
was observed that built up over ~40 minutes before quickly clearing.
This caused the verification stage to fail. An immediate reverify
succeeded as the spike had subsided.

To avoid this situation in future, we extend maas retries to 15 minutes,
and retry the whole verification phase 3 times. This will allow for the
duration of a similar spike, and time for the alarm status to change in
MaaS.