RLM-895 Remove old Kilo and Liberty jobs
This change removes the old style rpc-openstack jobs in rpc_aio.yml
that rarely pass and are covered elsewhere in rpc_openstack.yml using
the new gating methodology.

This change removes the Kilo/Liberty jobs from rpc_openstack.yml,
which are old, not maintained, and are somewhat redundant as we test
Kilo/Liberty installations as part of rpc_upgrades.yml.

This will reduce job duplication, and job failure reporting that the
triage team currently investigates for no good reason as Kilo/Liberty
are EOL.

This change is meant to remove:
- jenkins-PM_rpc-openstack-kilo-trusty-ceph-deploy
- jenkins-PM_rpc-openstack-kilo-trusty-swift-deploy
- jenkins-PM_rpc-openstack-liberty-trusty-ceph-deploy
- jenkins-PM_rpc-openstack-liberty-trusty-swift-deploy