RLM-316 Add PM deploy tests for artifact 'loose' mode
The 'loose' mode of using artifacts has been introduced
to help with initial deployments where the packages on
the hosts may be newer than those in the apt artifacts
repo. This patch adds two more tests which facilitate
verifying that this mode works on a day to day basis.

The existing tests are not changed so that the jobs
retain their history and so that the merge trigger
jobs also do not need to change.