RE-1191 Remove usages of the ansicolor plugin
This plugin uses Jenkins console notes to store color information,
this is hugely inefficient as it involves storing an encoded java
object in every line.

An alternative colourisation approach is to colourise on the client side
using javascript. This has already been implemented, so we can now
remove the server side implementation.