RE-1150 Remove RPC-O MNAIO jobs, settings and scripts
As discussed at length in RE-1150 and the early version
of this patch [1], the hardening and RPC-O jobs using
the old MNAIO scripts are not working, and are not
maintained and not monitored.

In order to reduce tech debt these are being removed. The
multi_node_aio_maas.* playbooks are deliberately left in
place to allow all MaaS entities to be cleaned up once the
jobs are removed. The legacy playbooks will be removed once
this is confirmed.

The hardening job is being replaced in RE-1238 and the
MNAIO jobs are being replaced with a more modern approach
implemented using in-repo scripts.

[1] https://github.com/rcbops/rpc-gating/pull/561